REFACTOR:Reorganizing hooks and types into subfolders by sense.
