+++ src/api/blogs/blogsApi.ts
@@ -1,6 +1,6 @@
 import { baseApi } from "../baseApi";
 
-import type { BlogDataTypes } from "../../types/blogDataTypes";
+import type { BlogDataTypes } from "../../types/blog/blogDataTypes";
 
 export const blogsApi = baseApi.injectEndpoints({
     endpoints: (create) => ({
