+++ src/components/organisms/pages/shopPage/SaleProducts/SaleProducts.tsx
@@ -1,4 +1,4 @@
-import useApiResource from '../../../../../hooks/useApiResource';
+import useApiResource from '../../../../../hooks/api/useApiResource';
 
 import { useGetProductsQuery } from '../../../../../api/products/productApi';
 
@@ -6,7 +6,7 @@ import SaleProductCard from '../../../../molecules/card/SaleProductCard/SaleProd
 
 import styles from '../SaleProducts/SaleProducts.module.scss';
 
-import type { ProductDataType } from '../../../../../types/productDataTypes';
+import type { ProductDataType } from '../../../../../types/product/productDataTypes';
 import { useMemo } from 'react';
 
 const SaleProducts:React.FC = () => {
