+++ src/components/organisms/pages/userDashboardPage/RecentOrderHistory/RecentOrderHistory.tsx
@@ -2,7 +2,7 @@ import { useMemo } from 'react';
 
 import { NavLink } from 'react-router-dom';
 
-import { useQueryDataByUserID } from '../../../../../hooks/useQueryDataByUserID';
+import { useQueryDataByUserID } from '../../../../../hooks/db/order/useQueryDataByUserID';
 
 import { selectOrdersStatus, selectOrderHistory, updateStatus, updateAllOrderHistory } from '../../../../../store/slices/orderHistorySlice';
