+++ src/hooks/products/useFilteredPromotedData.tsx
@@ -1,9 +1,9 @@
 import { useMemo } from "react";
 
 import { useGetProductsQuery } from "../../api/products/productApi";
-import useApiResource from "../useApiResource";
+import useApiResource from "../api/useApiResource";
 
-import { ProductDataType } from "../../types/productDataTypes";
+import { ProductDataType } from "../../types/product/productDataTypes";
 
 import type { FilterPromotedTypes as FilterType } from "../../types/filterPromotedType";
