+++ src/store/slices/mainSearchSlice.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import type { ProductDataType } from '../../types/productDataTypes';
+import type { ProductDataType } from '../../types/product/productDataTypes';
 
 interface MainSearchState{
   query: string;
