+++ src/store/slices/productModalSlice.ts
@@ -1,7 +1,7 @@
 import { createSlice } from '@reduxjs/toolkit';
 
 import type {PayloadAction} from '@reduxjs/toolkit';
-import { ProductDataType } from "../../types/productDataTypes";
+import { ProductDataType } from "../../types/product/productDataTypes";
 
 interface ProductModalState {
     isShow:boolean
