+++ src/types/product/categoryProductValueTypes.ts
