+++ src/types/product/productDataTypes.ts
