+++ src/utils/filter/subFilters/subFilterDate.ts
@@ -1,7 +1,7 @@
 type TDate = 'newest'|'oldest';
 
-import type { ProductDataType } from "../../../types/productDataTypes";
-import type { BlogDataTypes } from "../../../types/blogDataTypes";
+import type { ProductDataType } from "../../../types/product/productDataTypes";
+import type { BlogDataTypes } from "../../../types/blog/blogDataTypes";
 
 type Data = ProductDataType[]|BlogDataTypes[]
